Collate.r:
1. the 'prop.table.mx.grped.rows' function now gives proportions
for groups defined by
a 'groupnameprefixes' parameter rather than a 'numgroups' parameter.
This means that rather than having groups defined by splitting the
rows into 'numgroups' number of groups, it instead defines the groups by
looking for certain prefixes (i.e the 'groupnameprefixes') in the column
names.

2. the 'percentages_flattened_mx' function has been modified to use the
new 'prop.table.mx.grped.rows' as described above.

MatrixStats.r:

1. Added new function 'summary.grp' (returns rows of weighted
summaries - one for each group-by category) 
2. Added new function
'summary_mx_grpby_cols' - applies 'summary.grp' to columns of a matrix,
returning each column's summary as an
element of a list. The element of the list
consists of a summary in a row (or multiple summary rows if the grpby
parameter is used - one row for each grpby category for the column).
3. Changed 'mx' to x' in help of table.grpby function
4. Added '#' library(Hmisc)' to help file of wtdtable_mx_cols

MatrixUtil.r:

1. Removed the "|| ncol(mx) == 1" code in the 'flatten_mx_to_row'
function in the following code (commented out here):
"
cnames <- if (row.names.first 
					#|| ncol(mx) == 1
										) 
										{
				trim(paste(rep(rnames,length(cnames)), sapply(cnames, rep,
length(rnames))))
			} else {
				trim(paste(sapply(cnames, rep, length(rnames)),
rep(rnames,length(cnames))))
			}

"
This removed the inconsistency whereby if the matrix had only 1 column,
the function used to return a flattened matrix whose column names
had the rownames first, even when the
'row.names.first' parameter was given as 'FALSE' (the default).

The deletion of  the "|| ncol(mx) == 1" line of code, also enables the
function to return consistent naming of the output columns, regardless
of the number of input columns (previously one column input matrices had
their output matrices named differently than matrices with more
columns).

This changed was agreed upon by the original creator of the function -
Oliver Mannion.

Added to the help for this function.