Modified 'prop.table.mx.grped.rows' function to take the names of the groups (which are also in a portion of the names of the columns of the flattened matrix), from a vector previously attributed to the matrix -rather than using regular expressions to 'lop off' the group names from the column names (which was hit and miss).
We also changed the 'cmatchFlattened' function (used in the
'label_flattened_mx' function to label the columns of the new flattened
matrix) to have 'grpingNames' as an attribute to enable this. 

We also changed the 'label_flattened_mx' function, so that the
aforementioned attribute is assigned to the flattened matrix
so that the grouping names can be accessed in the
'prop.table.mx.grped.rows' function.

Also have that the 'if statement' in 'prop.table.mx.grped.rows'
now has '||is.null(grpingNames)' aswell.
 
Additional modifications to 'cmatchFlattened':
- put 'grping.match <- regexpr(".*\\s", x.flat)' code into 'else'
section where it really belongs.

Also help in label_flattened_mx - now uses dict_demo.